refactor(contribution-drafts): extract TempoCuesEditor/Display + shared draftKinds lib
DRY pass on the v0.5.0 contribution-drafts surfaces. Three pieces were
duplicating behaviour across sender + recipient + admin:

- Tempo-cues row editor with note-symbol palette lived inline inside
  ComposeDraftBlock. Extracted to TempoCuesEditor.tsx. Now also usable
  from the recipient's Edit & accept form.
- Read-only tempo-cues list lived inline in both ComposeDraftBlock
  ("compose-block-tempo*") and PendingDraftCard ("pending-draft-tempo*").
  Extracted to TempoCuesDisplay.tsx with its own styles; both call sites
  now render <TempoCuesDisplay raw={...} />.
- KIND_LABEL / KIND_ARTICLE maps were defined in ComposeDraftBlock and
  re-defined in RequestsAdmin (with a title-case variant). Consolidated
  into src/lib/draftKinds.ts as KIND_LABEL + KIND_LABEL_TITLE + KIND_ARTICLE,
  plus shared tempoCuesToRows / rowsToTempoCues / tempoCueEntries helpers.
  Every surface imports from the single source.

Side effect: recipients can now edit tempo cues through Edit & accept,
which silently carried through untouched before (the edit form didn't
surface the jsonb at all).

User requested DRY discipline be held as a cross-project engineering rule;
saved as a feedback memory + added to user global CLAUDE.md.
